GUI for GCRsim module
Graphic User Interface for running and analyzing cosmic ray simulations using the GCR module and converting into a pixelated SCA image. Updated charge diffusion (CD) code and added MCMC sampling on the final output array that attempts to fit the same CD model used to generate it as a test of the new functionality and laying the groundwork for further analysis of the simulated data.